Cata\[Fly][A - Quest] Twilight Highlands 84-85 [Kick].xml
Fixed a small issue during quest "The Only Homes We Have"
CJ already made a comment that the NPCs can be slow bringing buckets of water. However, waiting about 1 second for water to appear isn't much. My Bot actually stopped because the NPCs were too slow

[UseItemTargetLocation-v1085(error)] ItemId(62508) is not in our backpack

The WaitTime was increased to 3 seconds and VariantTime decreased to 2.  This may cause a slightly longer quest completion time, but it's better than your bot stopping because of slow NPCs.

This looks to be a problem with the QuestBehavior itself and a ticket has been submitted about it. However, I kept the increase in wait times because it should be a good thing.

Also fixed a problem I created on a previous commit. Profile should no longer use FlyTo incorrectly because of a missing If statement.
<If Condition="HasQuest(28598) &amp;&amp; IsQuestCompleted(28598)">